modified some folders name to avoid ambiguity (now we have two "clip" classes
