+++ scripts2/reference/generate_images_from_disk_model.py
@@ -8,7 +8,7 @@ from tqdm import tqdm
 from stable_diffusion2.utils.utils import save_images
 from stable_diffusion2.utils.model import initialize_autoencoder, initialize_clip_embedder, initialize_unet
 from stable_diffusion2.model.vae.autoencoder import Autoencoder
-from stable_diffusion2.model.clip_embedder import CLIPTextEmbedder
+from stable_diffusion2.model.clip_text_embedder_embedder import CLIPTextEmbedder
 from stable_diffusion2.model.unet import UNetModel
 from cli_builder import CLI
