+++ stable_diffusion2/model/__init__.py
@@ -11,3 +11,4 @@ summary: >
 * [U-Net](unet.html) with [attention](unet_attention.html)
 * [CLIP embedder](clip_embedder.html).
 """
+from . import *
