--- stable_diffusion2/model/clip/__init__.py
@@ -1,5 +1,0 @@
-from .clip_embedder import CLIPTextEmbedder
-
-__all__ = [
-    'CLIPTextEmbedder',
-]
