+++ stable_diffusion2/model/clip_text_embedder/__init__.py
@@ -1,0 +1,5 @@
+from .clip_text_embedder import CLIPTextEmbedder
+
+__all__ = [
+    'CLIPTextEmbedder',
+]
